Added magic autoloader
If the current SQL format is pgsql8, for instance, then
referencing class format_xyz will load pgsql8_xyz, and
alias pgsql8_xyz as format_xyz. If the class is not defined,
then the autoloader tries to load sql99_xyz. If an unloaded
format class is explicitly referred to, like pgsql8_table,
the autoloader loads that class instead.